+++ src/components/navbar.js
@@ -1,7 +1,14 @@
 import React from 'react';
 import Logo from '../assets/logo/logo-with-text.png';
+import { useContext } from 'react';
+import { ThemeContext } from '../master/theme';
 
-const navbar = () => {
+const menuItems = ['Home', 'Gallery', 'Bio', 'Skills', 'Facts', 'Quotes'];
+
+const Navbar = () => {
+
+  const { theme } = useContext (ThemeContext);
+
   return (
     <>
       <div className="h-14 w-full bg-pink-200 bg-opacity-45 flex">
@@ -14,35 +21,17 @@ const navbar = () => {
             <ul className="flex md:hidden justify-end h-full w-full ff-goldman text-md font-light">
               <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent hover:text-slate-50">
                 <i class="z-10 fa-solid fa-bars text-2xl cursor-pointer"></i>
-                <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
+                <div className={`absolute inset-0 bg-${theme} transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top`}></div>
               </li>
             </ul>
             <div className='hidden md:block h-full w-full'>
               <ul className="flex justify-end h-full w-full ff-goldman text-md font-light">
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Home</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
+              {menuItems.map((item, idx) => (
+                <li key={idx} className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
+                  <span className="z-10 group-hover:text-slate-50">{item}</span>
+                  <div className={`absolute inset-0 bg-${theme} transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top`}></div>
                 </li>
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Gallery</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
-                </li>
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Bio</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
-                </li>
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Skills</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
-                </li>
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Facts</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
-                </li>
-                <li className="h-full w-20 flex items-center justify-center cursor-pointer relative group transition-all ease-in-out duration-300 bg-transparent">
-                  <span className="z-10 group-hover:text-slate-50">Quotes</span>
-                  <div className="absolute inset-0 bg-chisato transition-all duration-300 transform scale-y-0 group-hover:scale-y-100 origin-top"></div>
-                </li>
+              ))}
               </ul>
             </div>
         </div>
@@ -51,4 +40,4 @@ const navbar = () => {
   )
 }
 
-export default navbar;
+export default Navbar;